requirements.txt
customtkinter==5.2.2
numpy==2.2.4
openai_whisper==20240930
Pillow==9.5.0
Pillow==11.1.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pyserial==3.5
pyttsx3==2.98
torch==2.5.1+cu121
